data_ins_project/requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.7.2
asttokens==2.2.1
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
black==23.3.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
cachetools==5.2.1
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.0.1
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
Django==4.2.3
django-bootstrap-v5==1.0.11
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
executing==1.2.0
filelock==3.6.0
fonttools==4.29.1
fs==2.4.12
gast==0.5.2
google-api-core==2.11.0
google-api-python-client==2.73.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.58.0
gunicorn==20.1.0
html5lib==1.1
httplib2==0.21.0
idna==3.4
importlib-metadata==4.6.4
ipython==7.31.1
jedi==0.18.2
jeepney==0.7.1
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.6
more-itertools==8.10.0
mpmath==0.0.0
mypy-extensions==1.0.0
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.2
olefile==0.46
packaging==23.0
parso==0.8.3
pathspec==0.11.1
pbr==5.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip-autoremove==0.10.0
platformdirs==3.2.0
ply==3.11
prompt-toolkit==3.0.38
protobuf==4.21.12
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycups==2.0.1
Pygments==2.14.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==0.19.2
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.8.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.4
stack-data==0.6.2
sympy==1.9
systemd-python==234
testresources==2.0.1
texttable==1.6.4
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.7.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
unittest2==1.1.0
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.13.0+ds
wadllib==1.3.6
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.2.3
xdg==5
xkit==0.0.0
zipp==1.0.0
